# delivery/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Testing Swagger APIs.",
        "title": "TODO APIs",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "version": "1.0"
    },
    "paths": {
        "/ai/chat": {
            "post": {
                "description": "Have a conversation with the AI model",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "Chat with AI",
                "parameters": [
                    {
                        "description": "Chat message",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.ChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "AI response",
                        "schema": {
                            "$ref": "#/definitions/controllers.ChatResponse"
                        }
                    },
                    "406": {
                        "description": "Invalid input format",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "AI model error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/ai/recommend": {
            "post": {
                "description": "Get AI-generated recommendations based on input data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "Get recommendations",
                "parameters": [
                    {
                        "description": "Input data for recommendation",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/infrastructure.Data"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully generated recommendations",
                        "schema": {}
                    },
                    "406": {
                        "description": "Invalid input format",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "AI model error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/ai/recommend-content": {
            "post": {
                "description": "Get AI-generated content recommendations based on input data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "Get content recommendations",
                "parameters": [
                    {
                        "description": "Input data for recommendation",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/infrastructure.DataXContent"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully generated content recommendations",
                        "schema": {
                            "$ref": "#/definitions/controllers.RecommendationResponse"
                        }
                    },
                    "406": {
                        "description": "Invalid input format",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "AI model error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/ai/recommend-tags": {
            "post": {
                "description": "Get AI-generated tags recommendations based on input data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "Get tags recommendations",
                "parameters": [
                    {
                        "description": "Input data for recommendation",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/infrastructure.DataXTags"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully generated tags recommendations",
                        "schema": {
                            "$ref": "#/definitions/controllers.RecommendationResponse"
                        }
                    },
                    "406": {
                        "description": "Invalid input format",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "AI model error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/ai/recommend-title": {
            "post": {
                "description": "Get AI-generated title recommendations based on input data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "Get title recommendations",
                "parameters": [
                    {
                        "description": "Input data for recommendation",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/infrastructure.DataXTitle"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully generated title recommendations",
                        "schema": {
                            "$ref": "#/definitions/controllers.RecommendationResponse"
                        }
                    },
                    "406": {
                        "description": "Invalid input format",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "AI model error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/ai/summarize": {
            "post": {
                "description": "Generate AI-powered summary of input content",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "Summarize content",
                "parameters": [
                    {
                        "description": "Input data to summarize",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/infrastructure.Data"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Generated summary",
                        "schema": {
                            "$ref": "#/definitions/controllers.SummaryResponse"
                        }
                    },
                    "406": {
                        "description": "Invalid input format",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "AI model error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controllers.ChatRequest": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Hello, how are you?"
                }
            }
        },
        "controllers.ChatResponse": {
            "type": "object",
            "properties": {
                "response": {
                    "type": "string",
                    "example": "I'm doing well, thank you for asking!"
                }
            }
        },
        "controllers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid data format"
                }
            }
        },
        "controllers.RecommendationResponse": {
            "type": "object",
            "properties": {
                "contents": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "titles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "controllers.SummaryResponse": {
            "type": "object",
            "properties": {
                "summary": {
                    "type": "string",
                    "example": "This is a summary of the provided content."
                }
            }
        },
...
        "infrastructure.Data": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "infrastructure.DataXContent": {
            "type": "object",
            "properties": {
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "infrastructure.DataXTags": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "infrastructure.DataXTitle": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}